chore: add terraform version
